+++ CHANGELOG.md
@@ -129,6 +129,9 @@
   - Added example ([issue #57](https://github.com/PowerShell/xFailOverCluster/issues/57))
     - 1-ChangeClusterNetwork.ps1
   - Added links to examples from README.md.
+- Changes to xClusterQuorum
+  - Refactored the unit test for this resource to use stubs and increase coverage
+    (issue #77).
 
 ### 1.6.0.0